@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+
 using System.Text;
 using System.Threading.Tasks;
 
@@ -8,9 +9,9 @@ namespace Celebrity.Model.Entities.SubEntities
 {
     public class CelebrityData
     {
-        public string Name { get; set; }
-        public string Date_Of_Birth { get; set; }
-        public string Town { get; set; }
+        public string Name { get; set; } = string.Empty;
+        public string Date_Of_Birth { get; set; } = string.Empty;
+        public string Town { get; set; } = string.Empty;
         public List<string> Movies { get; set; }
     }
 }
